move: OfferFullResponseSchema
