+++ app/routers/offer_full_routers.py
@@ -1,53 +1,18 @@
-from decimal import Decimal
-from enum import Enum
-from typing import Optional, List, Dict
-
 from fastapi import Depends, APIRouter, HTTPException, status
-from pydantic import Field, condecimal, validator
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
+from sqlalchemy import select
 
-from models import OfferLinkModel, EmployeeModel, ServiceNameModel
-from models.service_model import CategoryInSchema
+from models import OfferLinkModel, ServiceNameModel
 from routers.consts import RouteSlug
 from models.database import get_session
-from models.offer_model import OfferLinkInSchema
 from routers.offer_routers import OfferFilter
-from schemas.base_schema import BasePydanticSchema
-from schemas.offer_schema import OfferLinkOptionalSchema
-from services.service_service import OfferLinkService
-from dependencies.offer_dependency import valid_patch_id, valid_patch_schema
+from schemas.offer_schema import OfferFullResponseSchema
 
 router_offer_full = APIRouter()
 OFFER_SERVICE = '/offer/full/'
 
 
-from sqlalchemy import select
-
-
-class _OfferFullSchema(OfferLinkInSchema):
-    service: ServiceNameModel = Field(alias='service_name')
-    price: Optional[condecimal(max_digits=7, decimal_places=2)] = None
-
-    @validator('price', always=True)
-    def set_price(cls, v: Optional[Decimal], values: dict) -> Decimal:
-        return v or (Decimal(values['service'].price * values['rate'])).normalize()
-
-
-class OfferFullResponseSchema(BasePydanticSchema):
-    employee: EmployeeModel
-    offers: List[_OfferFullSchema] = []
-    categories: Dict[int, CategoryInSchema] = {}
-
-    @classmethod
-    def build(cls, data_db: List[OfferLinkModel]) -> 'OfferFullResponseSchema':
-        obj = cls(employee=data_db[0].employee)
-        for i in data_db:
-            full_schema = _OfferFullSchema.from_orm(i)
-            if full_schema.service.category_id not in obj.categories:
-                obj.categories[full_schema.service.category_id] = i.service_name.category
-            obj.offers.append(full_schema)
-        return obj
 
 
 @router_offer_full.get(OFFER_SERVICE + RouteSlug.ifilter + RouteSlug.pk, response_model=OfferFullResponseSchema)
