+++ app/tests/test_offer_full.py
@@ -8,11 +8,10 @@ from sqlalchemy.orm import sessionmaker
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from models import OfferLinkModel
-from routers.offer_full_routers import OfferFullResponseSchema
 from tests.utils import url_reverse
 from tests.conftest import engine
 from models.offer_model import OfferLinkInSchema
-from schemas.offer_schema import OfferLinkOptionalSchema
+from schemas.offer_schema import OfferLinkOptionalSchema, OfferFullResponseSchema
 from routers.offer_routers import OfferFilter
